Moving added functionality from `Logic/Taxonomy.php` to `Logic/ConsoleOutput/Taxonomy.php`
